the flip now works
I don't fully understand why it works. I guess because I was
setting useState outside of the component, it was changing for
everyone.

Now the logic is handled internally. I still need to think about
this for a while.

I also think the restructiring made things easier to implement
and understand. It also solved a problem I was getting with the
keys (I ws going to look into this, but I was guessing it was the
lack of keys for the outer card div.)