destructured
So I destructured because it's more direct. I also tested using
Card({cardName, cardText, ...otherProps}), but I think at this
moment it adds unnecessary complexity. If it's needed, I'll
implement it in the future.

All this tiny tests are just because I'm rewatching some of
epic react fundamentals, and trying along over here.