used useState and useEffect to load cards, flaky results
It's still a work in progress, but I'm fetching the cards,
and using useEffect to deal with the promise, and useState
to deal with the cardContent state. I still need to
understand this better.

IMPORTANT:
The results are flaky. Sometimes it works and sometimes not,
I don't know why. But the problem seems to be in the fetch.
The data and the error seem to not be coming trough.

Other stuff:
Since I'm always fetching random cards, I don't need shuffle,
I believe I removed it in a prior commit.

And in the end I thought the handleContent was an unnecessary
step. So I'm dealing directly with the data from fetchCards.