fixed the bug, fetch is working
It was a really dumb mistake. At some point I tried to get map
to work with cardContent, but it didn't. There was probably
something else that was wrong.

So I thought I would need to get something like a for loop,
I ended up doing that, but using map... It was very contrived.

But the problem was a combination of that, with the fact that
I wasn't passing an array initially. Since I had only one item,
when it tried to iterate to get the info, it broke.